Julia_Leonardi_z3 - "Random cellar"
My first 'bigger' program for "Introduction to Computer Science" class at the faculty of Cognitive Science (University of Warsaw). It draws a random amount of cabinets with a random amount of shelves and a random amount of arbitrarily chosen objects on the shelf. The size of the 'cellar' is given by the function's parameters. This program can be still improved by making the amount of cabinets dependent on the width of the cellar (as for 08.05.2020). Made using the turtle library.